requirements.txt
PyGithub>=1.53
ruamel.yaml>=0.16.12
Flask>=1.1.2
Flask-RESTful>=0.3.8
PyYAML>=5.3.1
DeepDiff>=5.0.0
